feat: 테스트 네이밍을 한글로 통일하고 단위 테스트 추가
- 모든 테스트 함수명을 한글로 변경 (43개)
- Presentation 레이어 단위 테스트 추가 (MusicUiState, PlaylistsScreenState)
- Utils 및 UseCase 단위 테스트 추가
- 스크린샷 테스트 예시 코드 추가
- 테스트 커버리지 향상 및 팀 소통 효율성 개선

+++ app/src/main/java/com/sdu/composemusicplayer/MusicAppState.kt
@@ -54,7 +54,7 @@ class MusicAppState(
             .stateIn(
                 scope = coroutineScope,
                 started = SharingStarted.WhileSubscribed(
-                    AndroidConstants.Time.MILLIS_IN_SECOND * AndroidConstants.Misc.SUBSCRIPTION_TIMEOUT_SECONDS
+                    AndroidConstants.Time.MILLIS_IN_SECOND * AndroidConstants.Misc.SUBSCRIPTION_TIMEOUT_SECONDS.toLong()
                 ),
                 initialValue = false,
             )
